Update README to direct users to specific version of typings
See discussion at https://github.com/DefinitelyTyped/DefinitelyTyped/pull/26363 -- typings for react-pointable had to be updated for compatibility with React 16.4 (since all of DefinitelyTyped is, unfortunately, versioned in lockstep). The upshot is that the latest version of @types/react-pointable will not expose pointer events when used with @types/react below 16.4.